+++ CHANGELOG.md
@@ -11,6 +11,7 @@ https://github.com/nwnxee/unified/compare/build8193.16...HEAD
 - Events: added skippable Acquire events to ItemEvents
 - Events: added `ACTION_RESULT` to Feat/Skill/Lock events for use in the _AFTER
 - Tweaks: `NWNX_TWEAKS_HIDE_PLAYERS_ON_CHAR_LIST`
+- Tweaks: `NWNX_TWEAKS_FIX_ARMOR_DEX_BONUS_UNDER_ONE`
 
 ##### New Plugins
 N/A